refactor: refactor utils folder
