+++ package.json
@@ -25,7 +25,6 @@
     "babel-plugin-styled-components": "^1.10.0",
     "classnames": "^2.2.6",
     "gatsby": "^2.8.8",
-    "gatsby-link": "^2.1.1",
     "gatsby-plugin-canonical-urls": "^2.0.13",
     "gatsby-plugin-catch-links": "^2.0.15",
     "gatsby-plugin-netlify": "^2.0.17",
