+++ src/components/foundations/reset/styles/base.ts
@@ -1,5 +1,5 @@
 import { css } from 'styled-components';
-import { textSizes, colors } from 'utils';
+import { textSizes, colors } from 'utils/variables';
 
 const base = css`
   :root {
