+++ src/components/layout/Footer/FooterWrapper.tsx
@@ -1,4 +1,5 @@
-import { styledWrapper as styled, space } from 'utils';
+import styled from 'styled-components';
+import { space } from 'utils/variables';
 
 const FooterWrapper = styled('div')`
   margin-top: ${space.xxl}px;
