+++ src/pages/404.tsx
@@ -3,7 +3,7 @@ import { graphql, Link } from 'gatsby';
 import Helmet from 'react-helmet';
 
 import { Page, NotFoundWrapper } from 'components/layout/Page';
-import { styledWrapper as styled } from 'utils';
+import { styledWrapper as styled } from 'utils/primitives';
 import { SiteMetadata } from 'interfaces/gatsby';
 import { RouteComponentProps } from '@reach/router';
 import { Heading, Text } from 'components/foundations';
