+++ src/templates/home.tsx
@@ -11,9 +11,9 @@ import { DocsHeader } from 'components/docs/DocsHeader';
 import MarkdownContent from 'components/page/Markdown/MarkdownContent';
 
 import { MenuNode, Edge } from 'interfaces/nodes';
-import { renderAst } from 'utils';
 import { Footer, FooterWrapper } from 'components/layout/Footer';
 import IndexLayout from 'layouts';
+import renderAst from 'utils/renderAst';
 // import FooterWrapper from 'components/old-layout/FooterWrapper';
 // import Footer from 'components/old-layout/Footer';
