--- src/utils.ts
@@ -1,3 +1,0 @@
-export * from './utils/primitives';
-export * from './utils/variables';
-export * from './utils/renderAst';
